requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiofiles==24.1.0
    # via gradio
annotated-doc==0.0.4
    # via
    #   fastapi
    #   typer
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   gradio
    #   httpx
    #   openai
    #   starlette
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
autogen-agentchat==0.7.5
    # via pyautogen
autogen-core==0.7.5
    # via autogen-agentchat
brotli==1.2.0
    # via gradio
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   pinecone-client
    #   requests
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
defusedxml==0.7.1
    # via fpdf2
distro==1.9.0
    # via openai
exceptiongroup==1.3.1
    # via anyio
fastapi==0.135.3
    # via gradio
ffmpy==1.0.0
    # via gradio
filelock==3.25.2
    # via
    #   huggingface-hub
    #   torch
fonttools==4.62.1
    # via
    #   fpdf2
    #   matplotlib
fpdf2==2.8.7
    # via -r requirements.in
fsspec==2026.3.0
    # via
    #   gradio-client
    #   huggingface-hub
    #   torch
google-search-results==2.4.2
    # via -r requirements.in
gradio==6.11.0
    # via -r requirements.in
gradio-client==2.4.0
    # via
    #   gradio
    #   hf-gradio
groovy==0.1.2
    # via gradio
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-gradio==0.3.0
    # via gradio
hf-xet==1.4.3
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   safehttpx
    #   tavily-python
huggingface-hub==1.10.1
    # via
    #   gradio
    #   gradio-client
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.1
    # via opentelemetry-api
jinja2==3.1.6
    # via
    #   gradio
    #   torch
jiter==0.13.0
    # via openai
joblib==1.5.3
    # via scikit-learn
jsonref==1.1.0
    # via autogen-core
jsonschema==4.26.0
    # via -r requirements.in
jsonschema-specifications==2025.9.1
    # via jsonschema
kiwisolver==1.5.0
    # via matplotlib
lxml==6.0.3
    # via python-pptx
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   gradio
    #   jinja2
matplotlib==3.10.8
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
narwhals==2.19.0
    # via plotly
networkx==3.4.2
    # via torch
numpy==2.2.6
    # via
    #   contourpy
    #   gradio
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
openai==2.31.0
    # via -r requirements.in
opentelemetry-api==1.41.0
    # via autogen-core
orjson==3.11.8
    # via gradio
packaging==26.0
    # via
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   matplotlib
    #   plotly
    #   transformers
pandas==2.3.3
    # via gradio
pillow==12.2.0
    # via
    #   autogen-core
    #   fpdf2
    #   gradio
    #   matplotlib
    #   python-pptx
pinecone-client==6.0.0
    # via -r requirements.in
pinecone-plugin-interface==0.0.7
    # via pinecone-client
plotly==6.7.0
    # via -r requirements.in
protobuf==5.29.6
    # via autogen-core
pyautogen==0.10.0
    # via -r requirements.in
pydantic==2.12.5
    # via
    #   autogen-core
    #   fastapi
    #   gradio
    #   openai
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via -r requirements.in
pydub==0.25.1
    # via gradio
pygments==2.20.0
    # via rich
pyparsing==3.3.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
    #   pinecone-client
python-dotenv==1.2.2
    # via pydantic-settings
python-multipart==0.0.24
    # via gradio
python-pptx==1.0.2
    # via -r requirements.in
pytz==2026.1.post1
    # via
    #   gradio
    #   pandas
pyyaml==6.0.3
    # via
    #   gradio
    #   huggingface-hub
    #   transformers
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2026.4.4
    # via
    #   tiktoken
    #   transformers
requests==2.33.1
    # via
    #   google-search-results
    #   tavily-python
    #   tiktoken
rich==14.3.3
    # via typer
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
safehttpx==0.1.7
    # via gradio
safetensors==0.7.0
    # via transformers
scikit-learn==1.7.2
    # via sentence-transformers
scipy==1.15.3
    # via
    #   scikit-learn
    #   sentence-transformers
semantic-version==2.10.0
    # via gradio
sentence-transformers==5.4.0
    # via -r requirements.in
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via openai
starlette==1.0.0
    # via
    #   fastapi
    #   gradio
sympy==1.14.0
    # via torch
tavily-python==0.7.23
    # via -r requirements.in
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.12.0
    # via tavily-python
tokenizers==0.22.2
    # via transformers
tomlkit==0.13.3
    # via gradio
torch==2.11.0
    # via sentence-transformers
tqdm==4.67.3
    # via
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
transformers==5.5.3
    # via sentence-transformers
typer==0.24.1
    # via
    #   gradio
    #   hf-gradio
    #   huggingface-hub
    #   transformers
typing-extensions==4.15.0
    # via
    #   anyio
    #   autogen-core
    #   exceptiongroup
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   opentelemetry-api
    #   pinecone-client
    #   pydantic
    #   pydantic-core
    #   python-pptx
    #   referencing
    #   sentence-transformers
    #   starlette
    #   torch
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
tzdata==2026.1
    # via pandas
urllib3==2.6.3
    # via
    #   pinecone-client
    #   requests
uvicorn==0.44.0
    # via gradio
xlsxwriter==3.2.9
    # via python-pptx
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
